Renames header tests/WorkArea.cpp to tests/WorkArea.hpp
