+++ tests/msim/test_msim_ACTIONX.cpp
@@ -45,7 +45,7 @@
 
 #include <opm/output/eclipse/EclipseIO.hpp>
 
-#include <tests/WorkArea.cpp>
+#include <tests/WorkArea.hpp>
 
 using namespace Opm;
