+++ tests/msim/test_msim_EXIT.cpp
@@ -33,7 +33,7 @@
 #include <opm/input/eclipse/Schedule/Schedule.hpp>
 #include <opm/input/eclipse/EclipseState/SummaryConfig/SummaryConfig.hpp>
 
-#include <tests/WorkArea.cpp>
+#include <tests/WorkArea.hpp>
 #include <opm/msim/msim.hpp>
 
 namespace Opm {
