+++ tests/parser/InitConfigTest.cpp
@@ -28,7 +28,7 @@
 #include <opm/input/eclipse/EclipseState/InitConfig/InitConfig.hpp>
 #include <opm/input/eclipse/Units/Units.hpp>
 
-#include <tests/WorkArea.cpp>
+#include <tests/WorkArea.hpp>
 
 #include <fstream>
