+++ tests/test_ESmry.cpp
@@ -35,7 +35,7 @@
 #include <math.h>
 #include <stdio.h>
 #include <tuple>
-#include "tests/WorkArea.cpp"
+#include "tests/WorkArea.hpp"
 
 using Opm::EclIO::ESmry;
