+++ tests/test_ExtESmry.cpp
@@ -37,7 +37,7 @@
 #include <stdio.h>
 #include <tuple>
 
-#include "tests/WorkArea.cpp"
+#include "tests/WorkArea.hpp"
 
 
 using Opm::EclIO::ESmry;
